Rename ControllerDataset file
